refactor(wallet): move file content to owner's place
