+++ libs/solana-dapp/wallet/utils/get-balances/src/lib/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './get-mint-accounts';
 export * from './get-native-account';
 export * from './get-token-accounts';
-export * from './create-balance';
